// src/interfaces/CompletedBuilds.ts
export interface CompletedBuildsInterface {
  id: number;
  name: string;
  cpu: number;
  cpuCooler: number;
  motherboard: number;
  memory: number;
  storage: number;
  videoCard: number;
  case: number;
  powerSuply: number;
  monitor: number;
  price: number;
  image: string;
  selected: boolean;
}